Move Linux-specific devicemapper code into driver_linux.go
